refactor: unify frequencyInserter_BCCW.js into frequencyInserter.js, the latter not being needed anymore
frequencyInserter_BCCWJ.js will be deleted in next commit. keeping it in this commit for testing.

check which corpus is loaded and adapt accordingly, instead of having two files with lots of duplicated code for both corpuses
(which was only a temporary solution to quickly get the BCCWJ version running anyways)

global variable (terms) in terms_BCCWJ.js and innocent_terms_complete.js needed to be changed from const to var
  to be available in window object